Almost completed webpage
Fixed the problems with my page "going crazy" when the width was decreasing/increasing, more stabile looking, added a @media for the #navbar, making the elements easier to read and eventually click when the page gets smaller. May add social's icons but not sure yet. Checking the whole code tomorrow, testing it and sending it over.